Сделанные шаблоны и поиск, без тестов
